mpm: cleanup: move mpm funcs into buffer specific files
